requirements.txt
streamlit_oauth
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
numpy
oauthlib
pandas
pillow
requests
streamlit